fix: Email and phone is now required UI
